Move UniquePasswordPolicy to its own policies sub-app
