+++ authentik/policies/password/api.py
@@ -4,7 +4,7 @@ from rest_framework.viewsets import ModelViewSet
 
 from authentik.core.api.used_by import UsedByMixin
 from authentik.policies.api.policies import PolicySerializer
-from authentik.policies.password.models import PasswordPolicy, UniquePasswordPolicy
+from authentik.policies.password.models import PasswordPolicy
 
 
 class PasswordPolicySerializer(PolicySerializer):
@@ -34,27 +34,6 @@ class PasswordPolicyViewSet(UsedByMixin, ModelViewSet):
 
     queryset = PasswordPolicy.objects.all()
     serializer_class = PasswordPolicySerializer
-    filterset_fields = "__all__"
-    ordering = ["name"]
-    search_fields = ["name"]
-
-
-class UniquePasswordPolicySerializer(PolicySerializer):
-    """Password Uniqueness Policy Serializer"""
-
-    class Meta:
-        model = UniquePasswordPolicy
-        fields = PolicySerializer.Meta.fields + [
-            "password_field",
-            "num_historical_passwords",
-        ]
-
-
-class UniquePasswordPolicyViewSet(UsedByMixin, ModelViewSet):
-    """Password Uniqueness Policy Viewset"""
-
-    queryset = UniquePasswordPolicy.objects.all()
-    serializer_class = UniquePasswordPolicySerializer
     filterset_fields = "__all__"
     ordering = ["name"]
     search_fields = ["name"]
