+++ authentik/policies/unique_password/tests/test_flows.py
@@ -1,0 +1,7 @@
+"""Unique Password Policy flow tests"""
+
+from authentik.flows.tests import FlowTestCase
+
+
+class TestUniquePasswordPolicyFlowBinding(FlowTestCase):
+    pass
