+++ authentik/policies/unique_password/urls.py
@@ -1,0 +1,7 @@
+"""API URLs"""
+
+from authentik.policies.password.api import UniquePasswordPolicyViewSet
+
+api_urlpatterns = [
+    ("policies/unique-password", UniquePasswordPolicyViewSet),
+]
