+++ authentik/root/settings.py
@@ -80,6 +80,7 @@ TENANT_APPS = [
     "authentik.policies.geoip",
     "authentik.policies.password",
     "authentik.policies.reputation",
+    "authentik.policies.unique_password",
     "authentik.policies",
     "authentik.providers.ldap",
     "authentik.providers.oauth2",
